Add import script for Friends+Me Google+ Exporter Community exports
Friends+Me Google+ Exporter ("F+MG+E") is the main tool currently available to
export an entire Google+ community.  This script takes the JSON output of one
or more Google+ communities, as well as the CSV file with archive image file
mappings, and a JSON file that describes how to map the Google+ communities to
existing categories, sub-categories, and tags in Discourse, and then imports
the data.  It represents Google+ posts as topics, and Google+ comments as posts
under those topics.  It fabricates somewhat reasonable topic titles from the
initial content of Google+ posts.

For any particular set of posts (one community, or one set of posts from one
community), it is a one-shot tool. It will not update to add new comments as
new posts to an existing topic.  However, it can be run many times to add
different content (separate communities, or separate posts from a community)
to an existing Discourse instance.  F+MG+E has a maximum number of posts per
JSON file, so very large communities might require multiple JSON files for
the import.

Unlike most other Discourse import tools, this one expects that the Discourse
categories and sub-categories have already been created.  It will not create
new categories for the import.  It is perfectly reasonable and expected to
import new content into an existing Discourse instance.

This scripts imports Google IDs for users who have posted, commented, and or
been plus-referenced, as long as they have not deleted their Google+ accounts
and broken the link to their Google ID. If Google authentication is set up for
the Discourse instance, user merging is intended to give users ownership of
their content imported under the same account they used for Google+